# Speed optimization of ToList() for Range() and Repeat() operations