The `Range()` and `Repeat()` operations already have a speed optimization for its `ToList()` method. They simply create a `List<T>` and add the elements to it. Here's the [implementation for Range](https://github.com/dotnet/corefx/blob/81c8c68d81daad033251d643e230061f38a9ae3a/src/System.Linq/src/System/Linq/Range.SpeedOpt.cs#L31):

```csharp
public List<int> ToList()
{
   List<int> list = new List<int>(_end - _start);
   for (int cur = _start; cur != _end; cur++)
   {
       list.Add(cur);
   }

   return list;
}
```

Although the length is passed as the `capacity` for the new `List<T>`, the `Add()` method still performs version increment, buffer length validation and size increment. This is performed for every element, which is unnecessary and penalizing.

```csharp
public void Add(T item)
{
    _version++;
    T[] array = _items;
    int size = _size;
    if ((uint)size < (uint)array.Length)
    {
        _size = size + 1;
        array[size] = item;
    }
    else
    {
        AddWithResize(item);
    }
}
```

As an alternative, the `List<T>(IEnumerable<T>)` constructor can be used, which has the following implementation:

```csharp
public List(IEnumerable<T> collection)
{
    if (collection == null)
        ThrowHelper.ThrowArgumentNullException(ExceptionArgument.collection);

    if (collection is ICollection<T> c)
    {
        int count = c.Count;
        if (count == 0)
        {
            _items = s_emptyArray;
        }
        else
        {
            _items = new T[count];
            c.CopyTo(_items, 0);
            _size = count;
        }
    }
    else
    {
        _size = 0;
        _items = s_emptyArray;
        using (IEnumerator<T> en = collection.GetEnumerator())
        {
            while (en.MoveNext())
            {
                Add(en.Current);
            }
        }
    }
}
```

When `collection` is not a `ICollection<T>`, which would be the case, `Add()` is still called for each element. But, when it's a non-empty `ICollection<T>`, it simply allocates an array, calls `ICollection<T>.CopyTo()` and sets the size. Only once!

**This pull request takes advantage of this to improve the speed optimization of `ToList()` for `Range` and `Repeat`.**

It adds private minimal implementations of `ICollection<T>` where the `CopyTo()` implementation simply uses a `for` loop to set the elements. With no validations!

Here's the implementation for `Range`:

```csharp
public void CopyTo(int[] array, int _)
{
    unchecked
    {
        for(int index = 0; index < _count; index++)
        {
            array[index] = _start + index;
        }
    }
}
```

Here's the implementation for `Repeat`:

```csharp
public void CopyTo(TResult[] array, int _)
{
    if (_current is object) // not null
    {
        unchecked
        {
            for(int index = 0; index < _count; index++)
            {
                array[index] = _current;
            }
        }
    }
}
```

The only side-effect is that an instance of the `ICollection<T>` has to be created but it will be very short lived.