@Wraith2 2019-04-29
Can you put together a BenchmarkDotNet project that shows the performance of standard vs this version? It would be a good idea to include the MemoryDiagnoser so the effect of the additional allocation can be seen and considered as well.