@aalmada 2019-04-29
@Wraith2 Sorry but I'm having an issue. I compiled _corefx_ in _Release_ mode, cloned [dotnet/performance](https://github.com/dotnet/performance/) and added new benchmarks. I can run these benchmarks for _dotnetcore3.0_ but I can't figure out how to point to my custom build. I'm following the [documentation](https://github.com/dotnet/performance/blob/master/docs/benchmarkdotnet.md#private-runtime-builds) but I can't find a _CoreRun.exe_ or a _dotnet.exe_. Do you know how it should be done?