@Wraith2 2019-04-29
The CI is having problems but they don't seem to be related to your work, results are [here](https://mc.dot.net/#/user/dotnet-bot/pr~2Fdotnet~2Fcorefx~2Frefs~2Fpull~2F37263~2Fmerge/test~2Ffunctional~2Fcli~2F/20190429.1) (since they can be hard to find). 

I'm not too familiar with the dotnet/performance repo but the instructions don't seem to cover the scenario of working on a local corefx build. That topic is covered in the corefx docs though. I'm sure that @adamsitnik will tell us that the easy way it to tell BDN where to find the corerun which is used the corefx build and test process, docs on that are here https://github.com/dotnet/coreclr/blob/master/Documentation/workflow/UsingCoreRun.md and the docs on how to tell BDN where to find it are in the BDN docs.

The other way is to setup a nuget.config file in your project directory (or above, see nuget.config docs) which clears all other nuget sources and then points at the `artifacts\packages\<build>` directory. I do this but you will end up having to edit the nuget package files to force in release binaries due to peculiarities of the package build process. This is time consuming but gets you a more standard dev experience with your local build.

The slightly cheaty way is that since you're doing something that I'd consider top level, I.e. nothing is resting on top of your changes, you could just put both implementations into an isolated test project side by side and see what number you get. It'll give reasonable comparative numbers as long as you're on a recent preview build already.