@ahsonkhan 2019-04-30
> I can run these benchmarks for _dotnetcore3.0_ but I can't figure out how to point to my custom build. I'm following the [documentation](https://github.com/dotnet/performance/blob/master/docs/benchmarkdotnet.md#private-runtime-builds) but I can't find a _CoreRun.exe_ or a _dotnet.exe_. Do you know how it should be done?

This should help (and if it doesn't, we should fix it): https://github.com/dotnet/performance/blob/master/docs/benchmarking-workflow-corefx.md

In the past, I have been able to run the benchmarks from dotnet/performance as follows:
```text
..\..\..\tools\dotnet\x64\dotnet.exe  run -c Release -f netcoreapp3.0 --coreRun "E:\GitHub\Fork\corefx\artifacts\bin\runtime\netcoreapp-Windows_NT-Release-x64\CoreRun.exe" --filter <MyBenchmarkClassName>
```