@aalmada 2019-04-30
@Wraith2 The CI issues seem to be related to #35117

I forgot to mention I'm running builds and benchmarks on MacOS but I finally got it working.

These are the benchmarks for multiple `count` values (0, 1, 10, 100, 1_000, 1_000_000). The first line is for `master` branch and the second for this PR branch. 

``` ini

BenchmarkDotNet=v0.11.3.1003-nightly, OS=macOS Mojave 10.14.4 (18E226) [Darwin 18.5.0]
Intel Core i5-7360U CPU 2.30GHz (Kaby Lake), 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=3.0.100-preview4-011223
  [Host]     : .NET Core 3.0.0-preview4-27615-11 (CoreCLR 4.6.27615.73, CoreFX 4.700.19.21213), 64bit RyuJIT
  Job-IYDZUB : .NET Core 36597c1f-8bf3-428e-a066-bcd32270d5b4 (CoreCLR 4.6.27624.71, CoreFX 4.700.19.23001), 64bit RyuJIT
  Job-YQBAIE : .NET Core 4efae4f4-4469-419f-80c8-932dcd3ca868 (CoreCLR 4.6.27624.71, CoreFX 4.700.19.23001), 64bit RyuJIT

Runtime=Core  IterationTime=250.0000 ms  MaxIterationCount=20  
MinIterationCount=15  WarmupCount=1  
```

## `Range(0, Count).ToList()`

| Method |                                                                                                     Toolchain |   Count |            Mean |           Error |          StdDev |          Median |             Min |             Max | Ratio | MannWhitney(10%) | RatioSD | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------- |-------------------------------------------------------------------------------------------------------------- |-------- |----------------:|----------------:|----------------:|----------------:|----------------:|----------------:|------:|----------------- |--------:|------------:|------------:|------------:|--------------------:|
| **ToList** |          **/corefx/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun** |       **0** |        **20.92 ns** |       **0.7615 ns** |       **0.8148 ns** |        **20.97 ns** |        **19.58 ns** |        **22.47 ns** |  **1.06** |             **Same** |    **0.05** |      **0.1958** |           **-** |           **-** |                **32 B** |
| ToList | /corefx_upstream/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun |       0 |        19.72 ns |       0.6978 ns |       0.8036 ns |        19.38 ns |        18.91 ns |        21.44 ns |  1.00 |             Base |    0.00 |      0.1958 |           - |           - |                32 B |
|        |                                                                                                               |         |                 |                 |                 |                 |                 |                 |       |                  |         |             |             |             |                     |
| **ToList** |          **/corefx/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun** |       **1** |        **56.12 ns** |       **3.3038 ns** |       **3.8046 ns** |        **54.80 ns** |        **52.06 ns** |        **63.64 ns** |  **1.46** |           **Slower** |    **0.10** |      **0.5386** |           **-** |           **-** |                **88 B** |
| ToList | /corefx_upstream/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun |       1 |        37.72 ns |       0.4806 ns |       0.4496 ns |        37.56 ns |        37.00 ns |        38.56 ns |  1.00 |             Base |    0.00 |      0.3917 |           - |           - |                64 B |
|        |                                                                                                               |         |                 |                 |                 |                 |                 |                 |       |                  |         |             |             |             |                     |
| **ToList** |          **/corefx/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun** |      **10** |        **65.65 ns** |       **0.3740 ns** |       **0.3123 ns** |        **65.66 ns** |        **65.05 ns** |        **66.27 ns** |  **1.17** |           **Slower** |    **0.02** |      **0.7343** |           **-** |           **-** |               **120 B** |
| ToList | /corefx_upstream/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun |      10 |        56.59 ns |       1.3781 ns |       1.2216 ns |        56.15 ns |        55.39 ns |        59.69 ns |  1.00 |             Base |    0.00 |      0.5873 |           - |           - |                96 B |
|        |                                                                                                               |         |                 |                 |                 |                 |                 |                 |       |                  |         |             |             |             |                     |
| **ToList** |          **/corefx/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun** |     **100** |       **209.17 ns** |       **1.9625 ns** |       **1.7397 ns** |       **208.38 ns** |       **207.69 ns** |       **212.90 ns** |  **0.75** |           **Faster** |    **0.01** |      **2.9325** |           **-** |           **-** |               **480 B** |
| ToList | /corefx_upstream/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun |     100 |       278.70 ns |       2.6635 ns |       2.3611 ns |       278.76 ns |       275.17 ns |       284.31 ns |  1.00 |             Base |    0.00 |      2.7849 |           - |           - |               456 B |
|        |                                                                                                               |         |                 |                 |                 |                 |                 |                 |       |                  |         |             |             |             |                     |
| **ToList** |          **/corefx/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun** |    **1000** |     **1,533.15 ns** |      **10.6542 ns** |       **8.8968 ns** |     **1,533.69 ns** |     **1,518.09 ns** |     **1,549.45 ns** |  **0.62** |           **Faster** |    **0.03** |     **24.3860** |           **-** |           **-** |              **4080 B** |
| ToList | /corefx_upstream/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun |    1000 |     2,480.10 ns |      97.4008 ns |     100.0235 ns |     2,490.28 ns |     2,382.13 ns |     2,777.22 ns |  1.00 |             Base |    0.00 |     24.3812 |           - |           - |              4056 B |
|        |                                                                                                               |         |                 |                 |                 |                 |                 |                 |       |                  |         |             |             |             |                     |
| **ToList** |          **/corefx/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun** | **1000000** | **1,824,031.94 ns** | **181,157.3112 ns** | **208,621.0289 ns** | **1,811,312.41 ns** | **1,529,766.74 ns** | **2,283,255.65 ns** |  **0.74** |           **Faster** |    **0.11** |    **298.6111** |    **291.6667** |    **291.6667** |           **4000080 B** |
| ToList | /corefx_upstream/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun | 1000000 | 2,487,097.92 ns | 113,476.7910 ns | 121,418.8950 ns | 2,505,794.86 ns | 2,200,668.02 ns | 2,728,511.64 ns |  1.00 |             Base |    0.00 |    270.8333 |    270.8333 |    270.8333 |           4000056 B |

## `Repeat<int>(1, Count).ToList()`

|        Method |                                                                                                     Toolchain |   Count |            Mean |           Error |          StdDev |          Median |             Min |             Max | Ratio | MannWhitney(10%) | RatioSD | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|-------------- |-------------------------------------------------------------------------------------------------------------- |-------- |----------------:|----------------:|----------------:|----------------:|----------------:|----------------:|------:|----------------- |--------:|------------:|------------:|------------:|--------------------:|
| **Repeat_ToList** |          **/corefx/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun** |       **0** |        **21.69 ns** |       **1.0962 ns** |       **1.2184 ns** |        **21.58 ns** |        **19.78 ns** |        **24.60 ns** |  **1.05** |             **Same** |    **0.08** |      **0.1958** |           **-** |           **-** |                **32 B** |
| Repeat_ToList | /corefx_upstream/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun |       0 |        20.54 ns |       0.7816 ns |       0.9001 ns |        20.64 ns |        19.26 ns |        22.40 ns |  1.00 |             Base |    0.00 |      0.1958 |           - |           - |                32 B |
|               |                                                                                                               |         |                 |                 |                 |                 |                 |                 |       |                  |         |             |             |             |                     |
| **Repeat_ToList** |          **/corefx/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun** |       **1** |        **54.05 ns** |       **1.1867 ns** |       **1.3190 ns** |        **53.72 ns** |        **52.50 ns** |        **57.01 ns** |  **1.46** |           **Slower** |    **0.03** |      **0.5385** |           **-** |           **-** |                **88 B** |
| Repeat_ToList | /corefx_upstream/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun |       1 |        37.02 ns |       0.3607 ns |       0.3012 ns |        36.99 ns |        36.58 ns |        37.68 ns |  1.00 |             Base |    0.00 |      0.3915 |           - |           - |                64 B |
|               |                                                                                                               |         |                 |                 |                 |                 |                 |                 |       |                  |         |             |             |             |                     |
| **Repeat_ToList** |          **/corefx/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun** |      **10** |        **64.52 ns** |       **1.1040 ns** |       **0.9787 ns** |        **64.13 ns** |        **63.55 ns** |        **66.92 ns** |  **1.17** |           **Slower** |    **0.02** |      **0.7342** |           **-** |           **-** |               **120 B** |
| Repeat_ToList | /corefx_upstream/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun |      10 |        55.14 ns |       0.5603 ns |       0.4678 ns |        54.98 ns |        54.70 ns |        56.39 ns |  1.00 |             Base |    0.00 |      0.5873 |           - |           - |                96 B |
|               |                                                                                                               |         |                 |                 |                 |                 |                 |                 |       |                  |         |             |             |             |                     |
| **Repeat_ToList** |          **/corefx/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun** |     **100** |       **198.45 ns** |       **1.8398 ns** |       **1.6309 ns** |       **198.20 ns** |       **196.05 ns** |       **202.34 ns** |  **0.68** |           **Faster** |    **0.01** |      **2.9319** |           **-** |           **-** |               **480 B** |
| Repeat_ToList | /corefx_upstream/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun |     100 |       290.87 ns |       4.5070 ns |       4.2158 ns |       288.91 ns |       287.25 ns |       302.26 ns |  1.00 |             Base |    0.00 |      2.7851 |           - |           - |               456 B |
|               |                                                                                                               |         |                 |                 |                 |                 |                 |                 |       |                  |         |             |             |             |                     |
| **Repeat_ToList** |          **/corefx/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun** |    **1000** |     **1,434.97 ns** |       **6.1796 ns** |       **4.8246 ns** |     **1,434.54 ns** |     **1,429.19 ns** |     **1,443.02 ns** |  **0.55** |           **Faster** |    **0.01** |     **24.3890** |           **-** |           **-** |              **4080 B** |
| Repeat_ToList | /corefx_upstream/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun |    1000 |     2,624.49 ns |      35.4596 ns |      31.4340 ns |     2,613.53 ns |     2,585.92 ns |     2,687.25 ns |  1.00 |             Base |    0.00 |     24.3825 |           - |           - |              4056 B |
|               |                                                                                                               |         |                 |                 |                 |                 |                 |                 |       |                  |         |             |             |             |                     |
| **Repeat_ToList** |          **/corefx/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun** | **1000000** | **1,671,097.37 ns** | **162,721.9699 ns** | **187,390.8625 ns** | **1,642,696.33 ns** | **1,429,412.55 ns** | **1,996,340.68 ns** |  **0.58** |           **Faster** |    **0.10** |    **289.7727** |    **284.0909** |    **284.0909** |           **4000080 B** |
| Repeat_ToList | /corefx_upstream/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun | 1000000 | 2,918,942.98 ns | 258,081.4497 ns | 297,206.9812 ns | 2,919,140.97 ns | 2,383,675.19 ns | 3,467,336.97 ns |  1.00 |             Base |    0.00 |    281.2500 |    281.2500 |    281.2500 |           4000056 B |

## Conclusions

- An extra 24 bytes are allocated.
- The new implementation is slower for small collections (down to 1.46 slower) but faster for large collections (up to 0.58 faster). 
- The performance is the same for empty collections.

One possible solution is to define a threshold and use one solution or the other, depending on the collection size.